Fix error in task 1 code

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -9,7 +9,7 @@ import asyncio
 from typing import AsyncGenerator
 
 
-async def async_generator() -> AsyncGenerator[float, None, None]:
+async def async_generator() -> AsyncGenerator[int, None, None]:
     """The coroutine will loop 10 times, each time asynchronously wait 1
     second, then yield a random number between 0 and 10.
     """
